+++ R/add_plot.R
@@ -18,11 +18,14 @@ ggplot_add.raster <- ggplot_add.grob
 #' @importFrom ggplot2 ggplot_add
 #' @export
 ggplot_add.nativeRaster <- ggplot_add.grob
+#' @importFrom ggplot2 ggplot_add
+#' @export
+ggplot_add.gt_tbl <- ggplot_add.grob
 
 #' @importFrom grid is.grob
 #' @importFrom grDevices is.raster
 should_autowrap <- function(x) {
-  is.grob(x) || inherits(x, 'formula') || is.raster(x) || inherits(x, 'nativeRaster')
+  is.grob(x) || inherits(x, 'formula') || is.raster(x) || inherits(x, 'nativeRaster') || inherits(x, 'gt_tbl')
 }
 
 # Convert a plot with a (possible) list of patches into a self-contained